replace old painter
